Add -allowdynamicpropsasstatic option to VBSP
By default, VBSP will remove prop_static entities that use physics or
dynamic models (unless they were compiled with the allowstatic option).
The wiki claims that this enforces consistency in the behaviour of the
models, but this should really be left up to the level designer.

By specifying the -allowdynamicpropsasstatic option, this restriction
can be lifted. This allows dynamic and physics models to be used with
prop_static entities (provided that the 'static' box is checked in the
model viewer's info tab). The advantage of this is that the prop has
less overhead and that it can cast lightmap shadows and receive baked
per-vertex lighting.